Variable refactoring dependent on cursor point
Summary:
By using (thing-at-point 'symbol), the refactoring command becomes
highly dependent on the placement of the point at the time the
refactoring command is called. By using 'sexp, the placement of the
point doesn't matter.

The tests were never catching this because it only shows up when
php-mode is actually enabled. The tests have been updated to cover this
situation. I have also expanded them to verify that we never break snake
case functionality as well.

Fixes #4